build(cmake): bump CommonLibSSE-NG pin
Advances the vendored submodule from the 7.2.0 release commit to
CommonLibSSE-NG's `ng` at 7.5.4 plus #332, which consolidates
RE::BSGraphics::BSShaderAccumulator into RE::BSShaderAccumulator as an
alias. That class was previously a separate declaration of the same
engine struct, so renderMode was a raw uint32_t there and the named-enum
comparisons needed a static_cast. It is the typed RENDER_MODE through
either spelling now, so the cast goes.

The pin points at #332's head commit; it moves to that PR's `ng` merge
commit once it lands, which does not change this diff.

+++ src/Features/Upscaling/PerfMode/MenuBridge.cpp
@@ -21,12 +21,10 @@ void PerfMode::UIPassDispatch_Hook::thunk(RE::BSGraphics::BSShaderAccumulator* s
 {
 	auto& perfMode = globals::features::upscaling.perfMode;
 
-	// Only intercept the VR world-space UI pass when hook is active. renderMode here is a
-	// raw uint32_t (this class is a legacy duplicate BSShaderAccumulator declaration), so
-	// the enum needs an explicit cast rather than comparing types directly.
+	// Only intercept the VR world-space UI pass when hook is active.
 	auto& rtData = shaderAccumulator->GetRuntimeData();
 	if (!perfMode.hookActive || !perfMode.fakeDSV ||
-		rtData.renderMode != static_cast<uint32_t>(RE::BSShaderAccumulator::RENDER_MODE::kVRWorldSpaceUIPass)) {
+		rtData.renderMode != RE::BSShaderAccumulator::RENDER_MODE::kVRWorldSpaceUIPass) {
 		func(shaderAccumulator, renderFlags);
 		return;
 	}
